requirements.txt
flask==1.1.0
gunicorn==20.0.4
requests==2.25.1
pandas==1.2.3
python-dotenv==0.16.0
bokeh==0.12.5